Alterar as IDs de int para long nas entidades

+++ Embrace.API/Infrastructure/Persistence/AcaoSolidaria.cs
@@ -2,7 +2,7 @@
 {
     public class AcaoSolidaria
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Nome { get; set; }
         public string TipoEvento { get; set; } // ex: Frio, Enchente
         public string Cidade { get; set; }
@@ -10,7 +10,7 @@
         public string Descricao { get; set; }
         public int MetaItens { get; set; }
 
-        public int OngId { get; set; }
+        public long OngId { get; set; }
         public Ong Ong { get; set; }
 
         public ICollection<Doacao> Doacoes { get; set; }
