+++ Embrace.API/Infrastructure/Persistence/VoluntarioAcao.cs
@@ -2,10 +2,10 @@
 {
     public class VoluntarioAcao
     {
-        public int VoluntarioId { get; set; }
+        public long VoluntarioId { get; set; }
         public Voluntario Voluntario { get; set; }
 
-        public int AcaoSolidariaId { get; set; }
+        public long AcaoSolidariaId { get; set; }
         public AcaoSolidaria AcaoSolidaria { get; set; }
     }
 }
